+++ NEWS.md
@@ -1,5 +1,7 @@
 # brandseyer2 0.0.1.9011
 
+* The filter language is available.
+
 # brandseyer2 0.0.1.9010
 
 * Added `with_tag_parents()` to provide parent information to tags.