+++ NEWS.md
@@ -3,6 +3,7 @@
 * The filter language is available.
 * `logs()` now shows a progress bar in interactive settings.
 * `logs_retrosent()` now makes account retrosend data available.
+* Non brandseye staff won't get content for tweets.
 
 # brandseyer2 0.0.1.9010